refactor(security): migrate schema with app isolation

+++ yak-security/src/main/java/io/yak/framework/security/common/dto/config/ConfigDTO.java
@@ -1,6 +1,6 @@
 package io.yak.framework.security.common.dto.config;
 public class ConfigDTO {
-  private Integer id;
+  private Long id;
   private String valueGroup;
   private String valueName;
   private String value;
@@ -8,7 +8,7 @@ public class ConfigDTO {
   private String memo;
   private String operator;
 
-  public Integer getId() { return this.id; }
+  public Long getId() { return this.id; }
 
   public String getValueGroup() { return this.valueGroup; }
 
@@ -22,7 +22,7 @@ public class ConfigDTO {
 
   public String getOperator() { return this.operator; }
 
-  public void setId(Integer id) { this.id = id; }
+  public void setId(Long id) { this.id = id; }
 
   public void setValueGroup(String valueGroup) { this.valueGroup = valueGroup; }
 
@@ -47,8 +47,8 @@ public class ConfigDTO {
     if (!other.canEqual(this)) {
       return false;
     }
-    Integer this$id = this.getId();
-    Integer other$id = other.getId();
+    Long this$id = this.getId();
+    Long other$id = other.getId();
     if (this$id == null ? other$id != null
                         : !((Object)this$id).equals(other$id)) {
       return false;
@@ -96,7 +96,7 @@ public class ConfigDTO {
   public int hashCode() {
     int PRIME = 59;
     int result = 1;
-    Integer $id = this.getId();
+    Long $id = this.getId();
     result = result * 59 + ($id == null ? 43 : ((Object)$id).hashCode());
     Integer $status = this.getStatus();
     result =
@@ -124,7 +124,7 @@ public class ConfigDTO {
 
   public ConfigDTO() {}
 
-  public ConfigDTO(Integer id, String valueGroup, String valueName,
+  public ConfigDTO(Long id, String valueGroup, String valueName,
                    String value, Integer status, String memo, String operator) {
     this.id = id;
     this.valueGroup = valueGroup;
